gcp.py
import os
import json
import requests
from google.oauth2 import service_account
from googleapiclient.discovery import build
import re
import base64

class GCPExtension:
def __init__(self):
credentials_path = os.environ.get('GOOGLE_APPLICATION_CREDENTIALS')
credentials = service_account.Credentials.from_service_account_file(
credentials_path, scopes=['https://www.googleapis.com/auth/cloud-platform'])
self.service = build('cloudresourcemanager', 'v1', credentials=credentials)

def handle_message(self, message):
calendar_pattern = r"^calendar\s+(\w+)\s+(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2})\s+(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2})$"
drive_pattern = r"^drive\s+(\w+)\s+(.*)$"
mail_pattern = r"^mail\s+(\w+)\s+(\w+)\s+(.*)$"

if re.match(calendar_pattern, message):
event_name, start_time, end_time = re.groups()
# Call the Google Calendar API to create an event with the given details
self.create_calendar_event(event_name, start_time, end_time)
elif re.match(drive_pattern, message):
file_name, file_contents = re.groups()
# Call the Google Drive API to upload a file with the given name and contents
self.upload_drive_file(file_name, file_contents)
elif re.match(mail_pattern, message):
recipient_email, subject, body = re.groups()
# Call the Gmail API to compose an email to the given recipient with the given subject and body
self.compose_gmail_email(recipient_email, subject, body)
else:
# Handle messages that don't match any known commands
return f"Unknown command: {message}"

def handle_message(self, message):
calendar_pattern = r"^calendar\s+(\w+)\s+(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2})\s+(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2})$"
drive_pattern = r"^drive\s+(\w+)\s+(.*)$"
mail_pattern = r"^mail\s+(\w+)\s+(\w+)\s+(.*)$"

if re.match(calendar_pattern, message):
event_name, start_time, end_time = re.groups()
# Call the Google Calendar API to create an event with the given details
self.create_calendar_event(event_name, start_time, end_time)
elif re.match(drive_pattern, message):
file_name, file_contents = re.groups()
# Call the Google Drive API to upload a file with the given name and contents
self.upload_drive_file(file_name, file_contents)
elif re.match(mail_pattern, message):
recipient_email, subject, body = re.groups()
# Call the Gmail API to compose an email to the given recipient with the given subject and body
self.compose_gmail_email(recipient_email, subject, body)
else:
# Handle messages that don't match any known commands
return f"Unknown command: {message}"

def create_calendar_event(self, event_name, start_time, end_time):
event_data = {
"summary": event_name,
"description": "",
"start": {"dateTime": start_time},
"end": {"dateTime": end_time}
}
response = self.service.events().insert(calendarId='primary', body=event_data).execute()
return response.get('id') # Return the ID of the newly created event
```
With the `create_calendar_event` method implemented, our next step is to tackle `upload_drive_file`. We will need to use the Google Drive API to upload a file. The API requires us to send a POST request to `https://www.googleapis.com/upload/drive/v3/files` with the file metadata and contents in the request body. Let's start by implementing the file metadata construction. We'll need to figure out how to handle file contents afterward.
```python
def upload_drive_file(self, file_name, file_contents):
file_metadata = {
"name": file_name,
"mimeType": "*/*" # Set MIME type to wildcard for automatic detection
}
# TO DO: Figure out how to handle file contents

def upload_drive_file(self, file_name, file_contents):
encoded_contents = base64.b64encode(bytes(file_contents, 'utf-8')).decode('utf-8')
file_metadata = {
"name": file_name,
"mimeType": "*/*", # Set MIME type to wildcard for automatic detection
}
# Construct the POST request body
body = {
'metadata': file_metadata,
'media': {'mimeType': file_metadata['mimeType'], 'body': encoded_contents}
}
# Make the POST request to upload the file
response = self.service.files().create(body=body).execute()
return response.get('id') # Return the ID of the newly uploaded file

